+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Documentation/Developer/Conceptual/Testing/IntegrationTesting.md
@@ -1,0 +1,13 @@
+# Integration Testing
+
+The TimeWarp Architecture use [Fixie](https://github.com/fixie/fixie) for the Testing Framework.
+
+Fixie allows for customization of the Test Project Lifecycle. Which is made of the Discovery Phase and 
+
+TimeWarp Architecture favors Integration Testing over Unit testing. We recommend only mocking code you don't have control over. If you have control over it then test it also.
+
+# Libraries
+
+[Fixie](https://github.com/fixie/fixie)
+[FakeItEasy](https://github.com/FakeItEasy/FakeItEasy)
+[Fluent Assertions](https://github.com/fluentassertions/fluentassertions)
