+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Source/TypeScript/tsconfig.json
@@ -65,5 +65,6 @@
     /* Advanced Options */
     "skipLibCheck": true /* Skip type checking of declaration files. */,
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
-  }
+  },
+  "exclude": ["node_modules"]
 }
