+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Tests/Client.Integration.Tests/Features/Application/ApplicationState_Clone_Tests.cs
@@ -9,7 +9,7 @@ namespace ApplicationState
   {
     private ApplicationState ApplicationState => Store.GetState<ApplicationState>();
 
-    public Clone_Should(ClientHost aWebAssemblyHost) : base(aWebAssemblyHost) { }
+    public Clone_Should(TestClientApplication aWebAssemblyHost) : base(aWebAssemblyHost) { }
 
     public void Clone()
     {
