+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Tests/Client.Integration.Tests/Features/Counter/CounterState_Clone_Tests.cs
@@ -9,7 +9,7 @@ namespace CounterState
   {
     private CounterState CounterState => Store.GetState<CounterState>();
 
-    public Clone_Should(ClientHost aWebAssemblyHost) : base(aWebAssemblyHost) { }
+    public Clone_Should(TestClientApplication aWebAssemblyHost) : base(aWebAssemblyHost) { }
 
     public void Clone()
     {
