+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Tests/Client.Integration.Tests/Features/EventStream/EventStreamState_Clone_Tests.cs
@@ -10,7 +10,7 @@ namespace EventStreamState
   {
     private EventStreamState EventStreamState => Store.GetState<EventStreamState>();
 
-    public Clone_Should(ClientHost aWebAssemblyHost) : base(aWebAssemblyHost) { }
+    public Clone_Should(TestClientApplication aWebAssemblyHost) : base(aWebAssemblyHost) { }
 
     public void Clone()
     {
