+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Tests/Client.Integration.Tests/Features/WeatherForecast/WeatherForecastState_Clone_Tests.cs
@@ -11,7 +11,7 @@ namespace WeatherForecastsState
   {
     private WeatherForecastsState WeatherForecastsState => Store.GetState<WeatherForecastsState>();
 
-    public Clone_Should(ClientHost aWebAssemblyHost) : base(aWebAssemblyHost) { }
+    public Clone_Should(TestClientApplication aWebAssemblyHost) : base(aWebAssemblyHost) { }
 
     public void Clone()
     {
