+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Tests/Client.Integration.Tests/Features/WeatherForecast/WeatherForecastState_FetchWeatherForecastsAction_Tests.cs
@@ -17,7 +17,7 @@ namespace WeatherForecastsState
 
     public FetchWeatherForecastsAction_Should
     (
-      ClientHost aClientHost,
+      TestClientApplication aClientHost,
       TimeWarpBlazorServerApplication aTimeWarpBlazorServerApplication
     )
       : base(aClientHost)
