+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Tests/Client.Integration.Tests/Infrastructure/ClientLoaderTestConfiguration.cs
@@ -4,9 +4,17 @@ namespace TimeWarp.Blazor.Client.Integration.Tests.Infrastructure
   using TimeWarp.Blazor.Features.ClientLoaders;
   using TimeWarp.Blazor.Testing;
 
+  /// <summary>
+  /// Replaces the default ClientLoaderConfiguration
+  /// To reduce the delay from 10 seconds to 10 milliseconds.
+  /// So tests run much faster.
+  /// </summary>
   [NotTest]
   public class ClientLoaderTestConfiguration : IClientLoaderConfiguration
   {
+    /// <summary>
+    /// Shortent the Delay for tests
+    /// </summary>
     public TimeSpan DelayTimeSpan => TimeSpan.FromMilliseconds(10);
   }
 }
