+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Tests/Client.Integration.Tests/Infrastructure/ClientTestingConvention.cs
@@ -23,7 +23,7 @@ namespace TimeWarp.Blazor.Client.Integration.Tests.Infrastructure
 
     public override void ConfigureApplications(ServiceCollection aServiceCollection)
     {
-      aServiceCollection.AddSingleton<ClientHost>();
+      aServiceCollection.AddSingleton<TestClientApplication>();
       base.ConfigureApplications(aServiceCollection);
     }
   }
