+++ Source/TimeWarp.Blazor.Template/templates/TimeWarp.Blazor/Tests/Client.Integration.Tests/Infrastructure/TestClientApplication.cs
@@ -6,32 +6,38 @@ namespace TimeWarp.Blazor.Client.Integration.Tests.Infrastructure
   using System;
   using TimeWarp.Blazor.Testing;
   using Microsoft.JSInterop;
+  using TimeWarp.Blazor.Features.ClientLoaders;
+  using MediatR;
+  using System.Threading.Tasks;
+  using System.Threading;
 
+  /// <summary>
+  /// Creates the ServiceProvider for the Client and configures it on construction
+  /// </summary>
   [NotTest]
-  public class ClientHost
+  public class TestClientApplication
   {
-    /// <summary>
-    /// Gets the service collection.
-    /// </summary>
+    private readonly ISender ScopedSender;
     public IServiceProvider ServiceProvider { get; }
 
-    public ClientHost(TimeWarpBlazorServerApplication aTimeWarpBlazorServerApplication)
+    public TestClientApplication(TimeWarpBlazorServerApplication aTimeWarpBlazorServerApplication)
     {
       var services = new ServiceCollection();
       // Need an HttpClient to talk to the Server side configured before calling AddBlazorState.
       services.AddSingleton(aTimeWarpBlazorServerApplication.HttpClient);
       ConfigureServices(services);
-
       ServiceProvider = services.BuildServiceProvider();
+      ScopedSender = new ScopedSender(ServiceProvider);
     }
 
-    private void ConfigureServices(IServiceCollection aServiceCollection)
+    private static void ConfigureServices(IServiceCollection aServiceCollection)
     {
       Program.ConfigureServices(aServiceCollection);
 
       // Theres is no JSRuntime in testing as we don't have an actual browser
-      IJSRuntime fakeJsRuntime = A.Fake<IJSRuntime>(); 
+      IJSRuntime fakeJsRuntime = A.Fake<IJSRuntime>();
       aServiceCollection.Replace(ServiceDescriptor.Scoped(_ => fakeJsRuntime));
+      aServiceCollection.Replace(ServiceDescriptor.Scoped<IClientLoaderConfiguration, ClientLoaderTestConfiguration>());
 
       // Could replace ICurrentUserService here with a logged in one for tests that need to have logged in user.
 
@@ -41,5 +47,15 @@ namespace TimeWarp.Blazor.Client.Integration.Tests.Infrastructure
 
       //aServiceCollection.Replace(ServiceDescriptor.Scoped(_ => fakeCurrentUserService));
     }
+
+    public Task<TResponse> Send<TResponse>
+    (
+      IRequest<TResponse> aRequest,
+      CancellationToken aCancellationToken = default
+    ) => ScopedSender.Send(aRequest, aCancellationToken);
+
+    public Task<object> Send(object aRequest, CancellationToken aCancellationToken = default) =>
+      ScopedSender.Send(aRequest, aCancellationToken);
+
   }
 }
